Anime Season Searching V1.2
Implemented the suggestions Ghassen suggested so ita not too much too scroll when making the call in Discord.

Just left the anime title, genres, and MAL url in the embedded message.

Also added the ability so the user can ask for more than 5 anime as an extra parameter... if this paramter is not used, then it defaults to 5 again.

Might still need some tweaks but we shall see... Right now the result is really compact and easier to scroll through